+++ src/demo/ts/ephox/alloy/demo/ToolbarsDemo.ts
@@ -1,15 +1,19 @@
-import { Arr, Option } from '@ephox/katamari';
+import { Arr, Result } from '@ephox/katamari';
 import { Class, Element } from '@ephox/sugar';
 import * as Attachment from 'ephox/alloy/api/system/Attachment';
 import * as Gui from 'ephox/alloy/api/system/Gui';
 import { Container } from 'ephox/alloy/api/ui/Container';
-import { SplitToolbar } from 'ephox/alloy/api/ui/SplitToolbar';
+import { SplitFloatingToolbar } from 'ephox/alloy/api/ui/SplitFloatingToolbar';
+import { SplitSlidingToolbar } from 'ephox/alloy/api/ui/SplitSlidingToolbar';
 import { Toolbar } from 'ephox/alloy/api/ui/Toolbar';
 import { ToolbarGroup } from 'ephox/alloy/api/ui/ToolbarGroup';
 import * as HtmlDisplay from 'ephox/alloy/demo/HtmlDisplay';
 
 import * as DemoRenders from './forms/DemoRenders';
 import { document, console, window } from '@ephox/dom-globals';
+import * as DemoSink from 'ephox/alloy/demo/DemoSink';
+import { LazySink } from 'ephox/alloy/api/component/CommonTypes';
+import * as Layout from 'ephox/alloy/positioning/layout/Layout';
 
 // tslint:disable:no-console
 
@@ -18,6 +22,13 @@ export default (): void => {
   const body = Element.fromDom(document.body);
   Class.add(gui.element(), 'gui-root-demo-container');
   Attachment.attachSystem(body, gui);
+
+  const sink = DemoSink.make();
+  gui.add(sink);
+
+  const lazySink: LazySink = (_) => {
+    return Result.value(sink);
+  };
 
   const groups = () => {
     return Arr.map([
@@ -117,9 +128,9 @@ export default (): void => {
 
   const subject2 = HtmlDisplay.section(
     gui,
-    'This toolbar has overflow behaviour that uses a more drawer',
-    SplitToolbar.sketch({
-      uid: 'demo-toolstrip',
+    'This toolbar has overflow behaviour that uses a sliding more drawer',
+    SplitSlidingToolbar.sketch({
+      uid: 'demo-toolstrip-sliding',
       dom: {
         tag: 'div'
       },
@@ -135,7 +146,7 @@ export default (): void => {
         }
       },
       components: [
-        SplitToolbar.parts().primary({
+        SplitSlidingToolbar.parts().primary({
           dom: {
             tag: 'div',
             styles: {
@@ -143,10 +154,10 @@ export default (): void => {
             }
           }
         }),
-        SplitToolbar.parts().overflow({
+        SplitSlidingToolbar.parts().overflow({
           dom: {
             tag: 'div',
-            styles: {
+              styles: {
               'display': 'flex',
               'flex-wrap': 'wrap'
             }
@@ -164,12 +175,74 @@ export default (): void => {
     })
   );
 
-  const splitToolbar = subject2;
+  const splitSlidingToolbar = subject2;
   const gps2 = Arr.map(groups(), ToolbarGroup.sketch);
   console.log('gps2', gps2);
-  SplitToolbar.setGroups(splitToolbar, gps2);
+  SplitSlidingToolbar.setGroups(splitSlidingToolbar, gps2);
+
+  const subject3 = HtmlDisplay.section(
+    gui,
+    'This toolbar has overflow behaviour that uses a floating more drawer',
+    SplitFloatingToolbar.sketch({
+      uid: 'demo-toolstrip-floating',
+      dom: {
+        tag: 'div'
+      },
+      lazySink,
+      getAnchor: (comp) => {
+        return {
+          anchor: 'hotspot',
+          hotspot: SplitFloatingToolbar.getMoreButton(comp).getOrDie(),
+          layouts: {
+            onRtl: () => [ Layout.southeast ],
+            onLtr: () => [ Layout.southwest ]
+          }
+        }
+      },
+      parts: {
+        'overflow-group': DemoRenders.toolbarGroup({
+          items: [ ]
+        }),
+        'overflow-button': {
+          dom: {
+            tag: 'button',
+            innerHtml: 'More'
+          }
+        },
+        'overflow': {
+          dom: {
+            tag: 'div',
+            styles: {
+              'display': 'flex',
+              'flex-wrap': 'wrap'
+            }
+          }
+        }
+      },
+      components: [
+        SplitFloatingToolbar.parts().primary({
+          dom: {
+            tag: 'div',
+            styles: {
+              display: 'flex'
+            }
+          }
+        })
+      ],
+
+      markers: {
+        overflowToggledClass: 'demo-more-button-toggled'
+      }
+    })
+  );
+
+  const splitFloatingToolbar = subject3;
+  const gps3 = Arr.map(groups(), ToolbarGroup.sketch);
+  console.log('gps3', gps3);
+  SplitFloatingToolbar.setGroups(splitFloatingToolbar, gps3);
 
   window.addEventListener('resize', () => {
-    SplitToolbar.refresh(splitToolbar);
+    SplitSlidingToolbar.refresh(splitSlidingToolbar);
+    SplitFloatingToolbar.refresh(splitFloatingToolbar);
   });
 };