+++ src/test/ts/browser/ui/toolbar/SplitSlidingToolbarTest.ts
@@ -1,21 +1,21 @@
 import { ApproxStructure, Assertions, Step } from '@ephox/agar';
 import { UnitTest } from '@ephox/bedrock';
-import { Arr, Option } from '@ephox/katamari';
+import { Arr } from '@ephox/katamari';
 import { Css } from '@ephox/sugar';
 import * as GuiFactory from 'ephox/alloy/api/component/GuiFactory';
 import { Button } from 'ephox/alloy/api/ui/Button';
-import { SplitToolbar } from 'ephox/alloy/api/ui/SplitToolbar';
+import { SplitSlidingToolbar } from 'ephox/alloy/api/ui/SplitSlidingToolbar';
 import * as GuiSetup from 'ephox/alloy/api/testhelpers/GuiSetup';
 import * as PhantomSkipper from 'ephox/alloy/test/PhantomSkipper';
 import * as TestPartialToolbarGroup from 'ephox/alloy/test/toolbar/TestPartialToolbarGroup';
 
-UnitTest.asynctest('SplitToolbarTest', (success, failure) => {
+UnitTest.asynctest('SplitSlidingToolbarTest', (success, failure) => {
   // Tests requiring 'flex' do not currently work on phantom. Use the remote to see how it is
   // viewed as an invalid value.
   if (PhantomSkipper.skip()) { return success(); }
 
   GuiSetup.setup((store, doc, body) => {
-    const pPrimary = SplitToolbar.parts().primary({
+    const pPrimary = SplitSlidingToolbar.parts().primary({
       dom: {
         tag: 'div',
         classes: [ 'test-toolbar-primary' ]
@@ -23,7 +23,7 @@ UnitTest.asynctest('SplitToolbarTest', (success, failure) => {
       shell: true
     });
 
-    const pOverflow = SplitToolbar.parts().overflow({
+    const pOverflow = SplitSlidingToolbar.parts().overflow({
       dom: {
         tag: 'div',
         classes: [ 'test-toolbar-overflow' ]
@@ -32,7 +32,7 @@ UnitTest.asynctest('SplitToolbarTest', (success, failure) => {
     });
 
     return GuiFactory.build(
-      SplitToolbar.sketch({
+      SplitSlidingToolbar.sketch({
         dom: {
           tag: 'div',
           classes: [ 'test-split-toolbar' ],
@@ -68,7 +68,6 @@ UnitTest.asynctest('SplitToolbarTest', (success, failure) => {
         }
       })
     );
-
   }, (doc, body, gui, component, store) => {
 
     const makeButton = (itemSpec) => {
@@ -83,7 +82,7 @@ UnitTest.asynctest('SplitToolbarTest', (success, failure) => {
     const sResetWidth = (px) => {
       return Step.sync(() => {
         Css.set(component.element(), 'width', px);
-        SplitToolbar.refresh(component);
+        SplitSlidingToolbar.refresh(component);
       });
     };
 
@@ -168,7 +167,8 @@ UnitTest.asynctest('SplitToolbarTest', (success, failure) => {
           { items: Arr.map([ { text: 'C' }, { text: 'D' } ], makeButton) },
           { items: Arr.map([ { text: 'E' }, { text: 'F' }, { text: 'G' } ], makeButton) }
         ]);
-        SplitToolbar.setGroups(component, groups);
+        SplitSlidingToolbar.setGroups(component, groups);
+        SplitSlidingToolbar.toggle(component);
       }),
 
       sAssertGroups('width=400px (1 +)', [ group1, oGroup ], [ group2, group3 ]),