+++ changelog.md
@@ -6,6 +6,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+# [4.12.0] - 2019-04-30
+
+### Changed
+- Broke the `SplitToolbar` component into separate self contained components, so it now has a separate component for floating (`SplitFloatingToolbar`) and sliding (`SplitSlidingToolbar`).
+
+### Fixed
+- Improved the `SplitToolbar` component accessibility for floating toolbars.
+
 # [4.11.18] - 2019-04-30
 ### Added
 - Added `AnchorOverrides` to all remaining Anchors and AnchorSpecs, as well as dropdowns